requirements.txt
curl_cffi==0.5.5
fake_useragent==1.1.3
python-dotenv==1.0.0
Telethon==1.26.1
wolframalpha==5.0
chromadb==0.4.20
Pillow
langchain_together
